athena/requirements.txt
flask==3.0.0
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
openai==1.6.1
python-dotenv==1.0.0
langchain==0.0.353
langchain-community==0.0.7
llama-index==0.9.24
langchain-core==0.1.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
tiktoken==0.5.2